+++ ICD.Common.Utils/IcdEnvironment.SimplSharp.cs
@@ -30,9 +30,36 @@ namespace ICD.Common.Utils
 			{
 				const CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET param =
 					CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET.GET_CURRENT_IP_ADDRESS;
-				const EthernetAdapterType type = EthernetAdapterType.EthernetLANAdapter;
-				short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(type);
-				yield return CrestronEthernetHelper.GetEthernetParameter(param, id);
+				const EthernetAdapterType primaryType = EthernetAdapterType.EthernetLANAdapter;
+				const EthernetAdapterType secondaryType = EthernetAdapterType.EthernetLAN2Adapter;
+
+				string address1 = null;
+
+				try
+				{
+					short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(primaryType);
+					address1 = CrestronEthernetHelper.GetEthernetParameter(param, id);
+				}
+				catch (ArgumentException)
+				{
+				}
+				
+				if (!string.IsNullOrEmpty(address1))
+					yield return address1;
+
+				string address2 = null;
+
+				try
+				{
+					short adapter2Type = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(secondaryType);
+					address2 = CrestronEthernetHelper.GetEthernetParameter(param, adapter2Type);
+				}
+				catch (ArgumentException)
+				{
+				}
+
+				if (!string.IsNullOrEmpty(address2))
+					yield return address2;
 			}
 		}
 
@@ -46,9 +73,104 @@ namespace ICD.Common.Utils
 			{
 				const CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET param =
 					CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET.GET_MAC_ADDRESS;
+				const EthernetAdapterType primaryType = EthernetAdapterType.EthernetLANAdapter;
+				const EthernetAdapterType secondaryType = EthernetAdapterType.EthernetLAN2Adapter;
+
+				string macAddress1 = null;
+
+				try
+				{
+					short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(primaryType);
+					macAddress1 = CrestronEthernetHelper.GetEthernetParameter(param, id);
+				}
+				catch (ArgumentException)
+				{
+				}
+
+				if (!string.IsNullOrEmpty(macAddress1))
+					yield return macAddress1;
+
+				string macAddress2 = null;
+
+				try
+				{
+					short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(secondaryType);
+					macAddress2 = CrestronEthernetHelper.GetEthernetParameter(param, id);
+				}
+				catch (ArgumentException)
+				{
+				}
+
+				if (!string.IsNullOrEmpty(macAddress2))
+					yield return macAddress2;
+			}
+		}
+
+		/// <summary>
+		/// Gets the dhcp status of the processor.
+		/// </summary>
+		[PublicAPI]
+		public static string DhcpStatus
+		{
+			get
+			{
+				const CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET param =
+					CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET.GET_CURRENT_DHCP_STATE;
 				const EthernetAdapterType type = EthernetAdapterType.EthernetLANAdapter;
-				short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(type);
-				yield return CrestronEthernetHelper.GetEthernetParameter(param, id);
+
+				try
+				{
+					short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(type);
+					return CrestronEthernetHelper.GetEthernetParameter(param, id);
+				}
+				catch (ArgumentException)
+				{
+					return null;
+				}
+			}
+		}
+
+		/// <summary>
+		/// Gets the hostname of the processor.
+		/// </summary>
+		[PublicAPI]
+		public static IEnumerable<string> Hostname
+		{
+			get
+			{
+				const CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET param =
+					CrestronEthernetHelper.ETHERNET_PARAMETER_TO_GET.GET_HOSTNAME;
+				const EthernetAdapterType primaryType = EthernetAdapterType.EthernetLANAdapter;
+				const EthernetAdapterType secondaryType = EthernetAdapterType.EthernetLAN2Adapter;
+
+				string address1 = null;
+
+				try
+				{
+					short id = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(primaryType);
+					address1 = CrestronEthernetHelper.GetEthernetParameter(param, id);
+				}
+				catch (ArgumentException)
+				{
+				}
+
+				if (!string.IsNullOrEmpty(address1))
+					yield return address1;
+
+				string address2 = null;
+
+				try
+				{
+
+					short adapter2Type = CrestronEthernetHelper.GetAdapterdIdForSpecifiedAdapterType(secondaryType);
+					address2 = CrestronEthernetHelper.GetEthernetParameter(param, adapter2Type);
+				}
+				catch (ArgumentException)
+				{
+				}
+
+				if (!string.IsNullOrEmpty(address2))
+					yield return address2;
 			}
 		}